+++ js/core.js
@@ -5,8 +5,8 @@
 // -----------------------------------------------------------------------------
 // Global Constants
 // -----------------------------------------------------------------------------
-const APP_VERSION = "v0.16.0.4";
-const APP_BANNER = "v0.16.0.4: New This Version: Full System Editor";
+const APP_VERSION = "v0.16.0.5";
+const APP_BANNER = "v0.16.0.5: New This Version: Full System Editor";
 
 // -----------------------------------------------------------------------------
 // Application State